+++ portal/apps/groupedtags/config.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class GroupedTagsConfig(AppConfig):
+    name = 'groupedtags'
+    verbose_name = 'grupos de tags'